feat: implement routine management with create, update, delete, and add exercise functionalities

+++ src/main/java/com/entrena/entrenaplatform/trainning/application/internal/commandservices/ExerciseCommandServiceImpl.java
@@ -3,39 +3,60 @@ package com.entrena.entrenaplatform.trainning.application.internal.commandservic
 import com.entrena.entrenaplatform.trainning.domain.model.aggregates.Exercise;
 import com.entrena.entrenaplatform.trainning.domain.model.commands.CreateExerciseCommand;
 import com.entrena.entrenaplatform.trainning.domain.model.commands.DeleteExerciseCommand;
+import com.entrena.entrenaplatform.trainning.domain.model.commands.UpdateExerciseCommand;
 import com.entrena.entrenaplatform.trainning.domain.services.ExerciseCommandService;
 import com.entrena.entrenaplatform.trainning.infrastructure.persistence.jpa.repositories.ExerciseRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 @Service
+@Transactional
 public class ExerciseCommandServiceImpl implements ExerciseCommandService {
 
     private final ExerciseRepository exerciseRepository;
 
-    public ExerciseCommandServiceImpl(ExerciseRepository exerciseRepository){
+    public ExerciseCommandServiceImpl(ExerciseRepository exerciseRepository) {
         this.exerciseRepository = exerciseRepository;
     }
 
     @Override
     public Long handle(CreateExerciseCommand command) {
-        //TODO: Add validations
+        if (exerciseRepository.existsByNameAndMuscleGroup(command.name(), command.muscleGroup())) {
+            throw new RuntimeException("Exercise with same name and muscle group already exists");
+        }
 
         var exercise = new Exercise(command);
-        try{
-            exerciseRepository.save(exercise);
-        }catch (Exception e){
+        try {
+            var savedExercise = exerciseRepository.save(exercise);
+            return savedExercise.getId();
+        } catch (Exception e) {
             throw new RuntimeException("Error saving exercise: " + e.getMessage(), e);
         }
-        return exercise.getId();
+    }
+
+    @Override
+    public void handle(UpdateExerciseCommand command) {
+        var exercise = exerciseRepository.findById(command.exerciseId())
+                .orElseThrow(() -> new RuntimeException("Exercise not found"));
+
+        try {
+            exercise.updateExercise(command);
+            exerciseRepository.save(exercise);
+        } catch (Exception e) {
+            throw new RuntimeException("Error updating exercise: " + e.getMessage(), e);
+        }
     }
 
     @Override
     public void handle(DeleteExerciseCommand command) {
-        //TODO: Add verification (if the exercise exists then delete
-        try{
-           exerciseRepository.deleteById(command.exerciseId());
-        }catch (Exception e){
-            throw new RuntimeException("Error deleting the exercise: " + e.getMessage(), e);
+        if (!exerciseRepository.existsById(command.exerciseId())) {
+            throw new RuntimeException("Exercise not found");
+        }
+
+        try {
+            exerciseRepository.deleteById(command.exerciseId());
+        } catch (Exception e) {
+            throw new RuntimeException("Error deleting exercise: " + e.getMessage(), e);
         }
     }
 }
